remove global visualizer object

+++ src/video.c
@@ -126,7 +126,7 @@ avi_stream_init(
 }
 
 size_t
-avi_stream_write_header(avi_stream *stream, void *ctx, int(*w)(uint8_t *buf, size_t size, void *ctx)) {
+avi_stream_write_header(avi_stream *stream, void *ctx, size_t(*w)(uint8_t *buf, size_t size, void *ctx)) {
     return w(stream->avi_header,326,ctx);
 }
